refactor: refactor components name
