--- components/pages/auth/components/index.tsx
@@ -1,1 +1,0 @@
-export { LoginPage } from './login';
